tests: drivers: build_all: ieee802154: remove mcr20a from spi.dtsi
This driver can't be compiled on all platforms as it's a NXP specific
driver (requires NXP HAL). Because spi.dtsi file is used by app.overlay
(and so enabled in multiple configs, including native_posix) it needs to
be removed. Note that MCR20A driver is already covered by the
ieee802154.build.mcr20a test case, which runs specifically on a NXP
board.

Signed-off-by: Gerard Marull-Paretas <gerard.marull@nordicsemi.no>