net: l2: ieee802154: introduce device roles
Introduces IEEE 802.15.4 device roles for use in later change sets.

In the near term this will be used for coordinator-role support in
tests. Then additional coordinator-specific features will be gradually
introduced without exposing them to the user API, yet. Once a reasonable
feature set has been collected for coordinators, it will be exposed in
public APIs for use in applications.

Signed-off-by: Florian Grandel <fgrandel@code-for-humans.de>