kconfig: Set `BOOTLOADER_SRAM_SIZE` default to 0
The current default value for `BOOTLOADER_SRAM_SIZE` is set to an
arbitrary-seeming value of `16`, which feels like a random magic
number.

Given that adding an offset in SRAM for a bootloader should always be
a conscious choice for a specific platform, this PR sets the default
to `0`. An appropriate value can be set at the SoC or board level,
overriding this default.

Signed-off-by: Kevin Townsend <kevin.townsend@linaro.org>